gooey-gui/py/pyproject.toml
[tool.poetry]
name = "gooey-gui"
version = "0.6.0"
description = ""
authors = ["Dev Aggarwal <devxpy@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "*"
uvicorn = { extras = ["standard"], version = "*" }
redis = "^4.5.1"
furl = "^2.1.3"
python-decouple = "^3.6"
loguru = "^0.7.2"

opencv-contrib-python = { version = "^4.7.0.72", optional = true }
numpy = { version = "^1.25.0", optional = true }

[tool.poetry.extras]
image = ["opencv-contrib-python", "numpy"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
